Update task registration and dependency in build.gradle.kts
Modified the way tasks are registered and the 'unzip' task no longer directly depends on the 'zip' task in the kotlin_dsl_module's build.gradle.kts file. This change is done to allow more flexibility when working with tasks, based on the suggestions by Cedric Champeau. Now, the 'unzip' task reads from the 'zip' task output, which may improve the build speed by moving the task execution to an earlier stage, if 'zip' task output is ready.